project/requirements.txt
gitpython==3.1.11
dill==0.3.2
numpyro==0.7.*
pyro-ppl==1.7.*
torch==1.9.*
matplotlib==3.4.*
funsor==0.4.*